Renames some classes and methods and fixes test
