+++ src/test/java/ChatClientTest.java
@@ -1,37 +1,45 @@
-import interfaces.SocketConnection;
+import fakes.FakeIO;
+import fakes.FakePrintStreamWriter;
+import fakes.FakeServerSocket;
+import fakes.FakeSocket;
 import org.junit.Test;
 
-import java.io.*;
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.io.PrintStream;
 import java.util.Arrays;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-public class ClientTest {
-    FakeIO fakeInput = new FakeIO(Arrays.asList("Priya", "quit"));
-    FakeSocket socket = new FakeSocket();
+public class ChatClientTest {
+    private FakeSocket socket = new FakeSocket();
     private ByteArrayOutputStream recordedOutput = new ByteArrayOutputStream();
     private PrintStream output = new PrintStream(recordedOutput);
 
     @Test
     public void getsInputStreamFromServer() {
+        FakeIO fakeInput = new FakeIO(Arrays.asList("Priya", "quit"));
         ChatClient client = new ChatClient(fakeInput, socket);
-        client.writeToServer();
+        client.writeOutToAndReadInFromClient();
         assertTrue(socket.getOutputStream);
     }
 
     @Test
     public void closesSocketAfterUserEntersQuit() {
+        FakeIO fakeInput = new FakeIO(Arrays.asList("Priya", "quit"));
         ChatClient client = new ChatClient(fakeInput, socket);
-        socket.closed = true;
-        client.writeToServer();
+        socket.closed = false;
+        client.writeOutToAndReadInFromClient();
         assertTrue(socket.closed);
     }
 
     @Test
     public void writesInputToOutputStream() {
+        FakeIO fakeInput = new FakeIO(Arrays.asList("Priya", "quit"));
         ChatClient client = new ChatClient(fakeInput, socket);
-        client.writeToServer();
+        client.writeOutToAndReadInFromClient();
         assertEquals(fakeInput.output, "You're connected on port 4444\n" +
                                        "Enter your name to register:\n" +
                                        "type quit to exit\n" +
@@ -44,12 +52,11 @@ public class ClientTest {
         FakeServerSocket serverSocket = new FakeServerSocket(new FakeSocket());
         UserIO console = createConsole("Priya\nquit\n");
         ChatClient client = new ChatClient(console, socket);
-        ChatServer server = new ChatServer(serverSocket, console);
-        FakePrintWriter printWriter = new FakePrintWriter(socket);
-        client.writeMessageToServerUntilQuit(fakeInput.getInput(), printWriter);
-        server.readInFromClient();
+        ChatServer server = new ChatServer(console, serverSocket);
+        FakePrintStreamWriter printWriter = new FakePrintStreamWriter(socket);
+        client.writeMessageToServerUntilQuit((console.getInput()), printWriter);
+        server.readInFromAndWriteOutToClient();
         assertEquals(recordedOutput.toString(), "Welcome Priya\n" +
-                                                "Welcome Priya\n" +
                                                 "Priya has now joined the chat room\n" +
                                                 "Bye!\n");
     }
@@ -61,38 +68,3 @@ public class ClientTest {
     }
 }
 
-class FakeSocket implements SocketConnection {
-
-    public boolean getOutputStream;
-    public boolean getInputStream;
-    public boolean closed;
-
-    public FakeSocket() {
-        this.getInputStream = false;
-        this.closed = false;
-    }
-
-    @Override
-    public int getPort() {
-        return 4444;
-    }
-
-    @Override
-    public InputStream getInputStream() {
-        getInputStream = true;
-        String inputFromServer = "Priya";
-        return new ByteArrayInputStream(inputFromServer.getBytes());
-    }
-
-    @Override
-    public OutputStream getOutputStream() {
-        getOutputStream = true;
-        OutputStream outputStream = new ByteArrayOutputStream() ;
-        return outputStream;
-    }
-
-    @Override
-    public void close() {
-        closed = true;
-    }
-}
